User(send/delete request) , Employee(approve/deny)
Added a form where the user can submit a verification request, view or delete pending request or see the reasoning behind a denied request.

Added buttons for approving/denying requests and textbox where the employee can explain their reasoning.

In addition I have split the data seed generator in to two parts. First is a generator for reference data, second is for person data. This is because in the old version upon generating data there would be a conflict since the persons would be given cities while cities are not yet generated.